[zircon][syscalls] Make zx_time_t and zx_duration_t signed types
This change is part of a hard transition and breaks the layers above.

zx_time_t and zx_duration_t are now int64_t. ZX_TIME_INFINITE is now
INT64_MAX.

A negative time value passed to a syscall is treated as any other time
value that's "in the past".

Test: CQ; new unit tests; built and ran topaz system on Eve

ZX-2100 #comment

+++ system/ulib/fdio/unistd.c
@@ -1934,7 +1934,7 @@ int ppoll(struct pollfd* fds, nfds_t n,
         zx_time_t tmo = ZX_TIME_INFINITE;
         // Check for overflows on every operation.
         if (timeout_ts && timeout_ts->tv_sec >= 0 && timeout_ts->tv_nsec >= 0 &&
-            (uint64_t)timeout_ts->tv_sec <= UINT64_MAX / ZX_SEC(1)) {
+            timeout_ts->tv_sec <= INT64_MAX / ZX_SEC(1)) {
             zx_duration_t seconds_duration = ZX_SEC(timeout_ts->tv_sec);
             zx_duration_t duration =
                 zx_duration_add_duration(seconds_duration, timeout_ts->tv_nsec);
